Changed gamma and kappa to floating point throught code to save memmory. Made zs, source_x, source_r and DlDs protected in Source class and accessable through accessor functions so they can be over loaded for a MutiSource class.
Worked a bit on MultiSourceAnaGalaxy class. This was renamed from previous
SourceAnaGalaxy class.

Added an InputParameters class that is not finished or used yet.

+++ AnalyticNSIE/nfw_lens.cpp
@@ -34,7 +34,7 @@ void alphaNFW(double *alpha,double *x,double Rtrunc,double mass,double r_scale
 	return ;
 }
 /// Convergence for an NFW halo
-double kappaNFW(double *x,double Rtrunc,double mass,double r_scale
+float kappaNFW(double *x,double Rtrunc,double mass,double r_scale
 		,double *center,double Sigma_crit){
 	double r;
 	double gfunction(double),ffunction(double);
@@ -55,7 +55,7 @@ double kappaNFW(double *x,double Rtrunc,double mass,double r_scale
 }
 
 /// Shear for and NFW halo. this might have a flaw in it
-void gammaNFW(double *gamma,double *x,double Rtrunc,double mass,double r_scale
+void gammaNFW(float *gamma,double *x,double Rtrunc,double mass,double r_scale
 		,double *center,double Sigma_crit){
 	double r,gt=0;
 	double g2function(double x);
