+++ TreeCode/quadTree.cpp
@@ -400,7 +400,7 @@ void QuadTree::rotate_coordinates(double **coord){
  *       4*pi*G*mass_scale to get the deflection angle caused by the plane lens.
  * */
 
-void QuadTree::force2D(double *ray,double *alpha,double *kappa,double *gamma,bool no_kappa){
+void QuadTree::force2D(double *ray,double *alpha,float *kappa,float *gamma,bool no_kappa){
 
   PosType xcm,ycm,rcm2cell,rcm2,tmp,boxsize2;
   IndexType i;
