+++ include/point.h
@@ -42,12 +42,12 @@ typedef struct Point{
   Boo in_image; // marks if point is in image
 
   // redundant information in image and source points
-  double kappa;        // surface density
-  double gamma[3];    // shear, third component is the rotation quantity that is only non-zero for multi-plane lensing
-  double dt;          // time delay
-  double invmag;     // inverse of magnification
+  float kappa;        // surface density
+  float gamma[3];    // shear, third component is the rotation quantity that is only non-zero for multi-plane lensing
+  float dt;          // time delay
+  float invmag;     // inverse of magnification
   double gridsize;   // the size of the most refined grid the point is in
-  double surface_brightness;  // the surface brightness at this points
+  float surface_brightness;  // the surface brightness at this points
 
   struct branchstruct *leaf;
