+++ include/quadTree.h
@@ -158,7 +158,7 @@ public:
 			);
 	virtual ~QuadTree();
 
-	void force2D(double *ray,double *alpha,double *kappa,double *gamma,bool no_kappa);
+	void force2D(double *ray,double *alpha,float *kappa,float *gamma,bool no_kappa);
 	void printParticlesInBranch(unsigned long number);
 	void printBranchs(int level = -1);
